Add a `mention_pattern` method to DiscordUser
As evidenced by current efforts to port the Chat module, this will be
a useful API method. Protocols may implement this as a pattern that
should be guaranteed to match any reference to the bot, be it a bare
name or a specific "mention" format, even if it may differ (like in
Discord).