Simplify phrase fetching, enforce phrase cooldown
Due to PRNG being imperfect, it's not unlikely that ZeroBot may pull the
same phrases either back to back, or not very far apart. A configurable
cooldown setting will remember the last `n` phrases per table, which
will be checked every time a phrase is fetched. If the fetched phrase is
in the cooldown queue, another phrase will be fetched instead. The
`fetch_phrase` coroutine was added as a wrapper around this
functionality.