8Ball: Support more natural phrase combinations
Some preludes are worded such that they presuppose an action or
non-action type answer, which can make for an awkward 8Ball
message when that assumption is not met. For example, an action prelude
ending with something like "...and reveals the answer" followed by an
action answer doesn't logically fit.

A new optional column `expects_action` restricts the type of answer
that follows a prelude. If true, the answer is restricted to action
answers, and vice-versa. If `NULL` (the default and usual behavior),
then no restriction is placed. Currently, this column is only meaningful
for **prelude** response types.