Ditch `shlex` in favor of our own shell split
Added `ZeroBot.util.shellish_split`, a shell split tailored to ZeroBot's
needs. `shlex` did too many things that ended up not being a good fit
for ZeroBot commands:
    - It naturally escapes all escaped characters, but some ZeroBot
      commands (like `quote add`) need to be able to interpret these on
      their own.
    - There was no way to escape only *some* sequences; it was all or
      nothing.
    - Single-quotes are treated as substrings as well, but apostrophes
      are very common in chat; having to escape them would result in
      a miserable and frustrating user experience.

It's very likely not the best or fastest implementation, but it will
only be run once per command, so it's not a huge deal.