Fix `quote quick` not using the current channel
The initial implementation's logic for `last_messages` was...flawed.
`quick` is supposed to be limited to the source channel (except for
`--id`), but we were pulling from the most recent cached message
throughout the entire context.

Like before, the current implentation is still Discord-specific for now,
and will be updated to use a protocol-agnostic interface once I've
fleshed it out.