Chat: Periodically set random "activities"
Some protocols support an "activity" as part of or in addition to
a user's status. For example, Discord supports several activity states
as part of its rich presence such as "Playing", "Listening to",
"Watching", and so on.

ZeroBot will randomly pick activities from the database on
a configurable interval. For all protocols that support this activity
(currently just Discord).