feat(counter): Change when/how names are displayed
The `$user` expansion now always expands to the person who triggered the
counter, whether implicitly or explicitly via `!count`. As
a consequence, there are now two new expansions: `$r_user` and `$r_all`.
The latter expands to an or-joined list of each user that the
counter is restricted to. The former expands to either `$user` or
`$r_all` depending on who triggered the counter.

Additionally, `!count info` is a special case, where `$user` expands to
the last user that triggered the counter.