feat(database): Move database schemas into `sql` files
Resolves #40. Having the schemas crammed into the python files was
pretty obnoxious. Instead, leverage `importlib.resources` and move them
to discrete `sql` files.

Additionally, added a `create_database` function to `ZeroBot.database`.
Core will now check for the existence of a database and only execute the
core schema if it didn't exist. The feature schemas are still run every
time they're registered for now, but this should probably be changed in
the future.