fix(chat): Deprecated interface for Discord.py tasks
The 2.0 version of dpy allowed specifying discrete times to run a task
and in doing so updated the constructor for `Loop`. Annoyingly, only the
decorator has optional parameters, and instead of opting for `None`,
a sentinel object is used. This object must be passed in order to not
raise an exception since `__init__` requires both discrete and relative
time parameters, but `Loop.change_interval` will raise when both are
passed and one group isn't `MISSING`.