feat: Add `zerobotdb` entry point
This entry point can be use to spin up a connection to a ZeroBot
database that is already set up with the required customizations, e.g.
functions, collations, etc. and then drop into an interactive
interpreter.

Pretty rough for now, but in the future I intend to expand this entry
point with subcommands to do various database-related tasks like working
with backups, running `VACUUM`, getting stats about the db, etc.