refactor(discord): Cache fetched DMChannel
It took some looking about, but it seems that the intended way to cache
DMChannels is to specifically call `create_dm`. This adds it to the
bot's cache as expected, and then `get_channel` works properly with
recipient and all. "Create" is a bit of a misnomer and actually is more
like a "create or return existing".