feat(magic8ball): Add Discord embed output formatting
Leverages a new prototype interface specific to the Discord protocol
module. I can build a response as both a plain text message as I have
been, as well as a stylized Discord embed. Both versions are wrapped
into the `ResponseProxy` object, which transparently sends off the
appropriate version based on the protocol that issued the command.

Resolves issue #45.