feat(context): Support direct messages
While a direct message implies *direct* communication with another user,
it's not much of a stretch to imagine that there's still some kind of
communication *channel* in place. Given that the definition of a ZeroBot
Channel is an abstract medium to communicate with another user on, I say
that it fits the bill. Thus "direct messages" are simply a Channel with
two users.